+++ src/pileup.hpp
@@ -129,7 +129,7 @@ public:
     /// read from protobuf
     void load(istream& in);
     /// write to protobuf
-    void write(ostream& out, uint64_t buffer_size = 5);
+    void write(ostream& out, size_t buffer_size = 5);
 
     /// apply function to each pileup in table
     void for_each_node_pileup(const function<void(NodePileup&)>& lambda);
