+++ src/subcommand/align_main.cpp
@@ -211,8 +211,8 @@ int main_align(int argc, char** argv) {
     if (output_json) {
         cout << pb2json(alignment) << endl;
     } else {
-        function<Alignment(uint64_t)> lambda =
-            [&alignment] (uint64_t n) {
+        function<Alignment(size_t)> lambda =
+            [&alignment] (size_t n) {
                 return alignment;
             };
         stream::write(cout, 1, lambda);
